Added support for SR and UA descriptors into descriptor storage

+++ PathFinder/Source/Foundation/NameRegistry.hpp
@@ -9,18 +9,18 @@ namespace Foundation
     class NameRegistry
     {
     public:
-        static const uint64_t INVALID_ID = UINT64_MAX;
+        static const uint16_t INVALID_ID = UINT16_MAX;
 
         static NameRegistry& SharedInstance();
 
         NameRegistry();
         ~NameRegistry();
 
-        uint64_t ToId(const std::string& string);
-        const std::string& ToString(uint64_t id);
+        uint16_t ToId(const std::string& string);
+        const std::string& ToString(uint16_t id);
 
     private:
-        std::unordered_map<std::string, uint64_t> m_NameToId;
+        std::unordered_map<std::string, uint16_t> m_NameToId;
         std::vector<std::string> m_IdToName;
     };
 }
